Moving the useEffect enquiing the snackbar for errorMessages to a global Provider
