+++ packages/twenty-front/src/modules/auth/components/VerifyLoginTokenEffect.tsx
@@ -5,8 +5,6 @@ import { useIsLogged } from '@/auth/hooks/useIsLogged';
 import { useVerifyLogin } from '@/auth/hooks/useVerifyLogin';
 import { clientConfigApiStatusState } from '@/client-config/states/clientConfigApiStatusState';
 import { AppPath } from '@/types/AppPath';
-import { SnackBarVariant } from '@/ui/feedback/snack-bar-manager/components/SnackBar';
-import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
 import { useRecoilValue } from 'recoil';
 import { isDefined } from 'twenty-shared/utils';
 import { useNavigateApp } from '~/hooks/useNavigateApp';
@@ -14,9 +12,6 @@ import { useNavigateApp } from '~/hooks/useNavigateApp';
 export const VerifyLoginTokenEffect = () => {
   const [searchParams] = useSearchParams();
   const loginToken = searchParams.get('loginToken');
-  const errorMessage = searchParams.get('errorMessage');
-
-  const { enqueueSnackBar } = useSnackBar();
   const isLogged = useIsLogged();
   const navigate = useNavigateApp();
   const { verifyLoginToken } = useVerifyLogin();
@@ -26,13 +21,6 @@ export const VerifyLoginTokenEffect = () => {
   );
 
   useEffect(() => {
-    if (isDefined(errorMessage)) {
-      enqueueSnackBar(errorMessage, {
-        dedupeKey: 'get-auth-tokens-from-login-token-failed-dedupe-key',
-        variant: SnackBarVariant.Error,
-      });
-    }
-
     if (!clientConfigLoaded) return;
 
     if (isDefined(loginToken)) {
